rename Duke and give it ability to greet
