+++ src/main/java/Chunky.java
@@ -1,0 +1,17 @@
+package main.java;
+
+public class Chunky {
+    public static void main(String[] args) {
+        String logo = " ____        _        \n"
+                + "|  _ \\ _   _| | _____ \n"
+                + "| | | | | | | |/ / _ \\\n"
+                + "| |_| | |_| |   <  __/\n"
+                + "|____/ \\__,_|_|\\_\\___|\n";
+        System.out.println("----------------------------\n");
+        System.out.println("Hello! I'm Chunky!\n" + logo);
+        System.out.println("What can I do for you ?\n");
+        System.out.println("----------------------------\n");
+        System.out.println("Bye! Hope to see you again soon!");
+        System.out.println("----------------------------\n");
+    }
+}
